tests/wpt/web-platform-tests/cookies/path/set-from-http.sub.html

<!doctype html>
<html>
<head>
  <meta charset=utf-8>
  <title>Sends cookie from `Set-Cookie` HTTP header for path=/</title>
  <meta name=help href="http://tools.ietf.org/html/rfc6265#section-5.1.4">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/cookies/resources/testharness-helpers.js"></script>
</head>
<body>
<div id=log></div>
<script>
  function clearKnownCookie() {
    document.cookie = "secure_from_secure_http=0; Secure; expires=Thu, 01 Jan 1970 00:00:01 GMT; path=/";
  }

  test(function () {
    assert_equals(document.cookie.match(/Path=1/), null);
  }, "'Path=/' not present in `document.cookie`");

  promise_test(function (t) {
    t.add_cleanup(clearKnownCookie);
    return fetch("http://{{host}}:{{ports[http][0]}}/cookies/resources/echo-json.py",
                 { "credentials": "include" })
      .then(function (r) {
        return r.json();
      })
      .then(function (j) {
        assert_equals(j["path"], "secure_from_secure_http=1");
      });
  }, "'secure' cookie sent in HTTP request");
</script>
</body>
</html>
